[BUGFIX] : Fix displayed FIAT value when Hidding tokens in Portfolio
